Store authenticated TLS Client Cert DN in Event header
If a TLS connection has resulted in an authenticated TLS Client
Cert (either from client-auth "want" or "need"), then store the
certificate's Distinguished Name (DN) in a Flume event header called
"ssl-dn".

Any other value in this header will be replaced, and any other use
of this header (for a non-authenticated client, or non-SSL) will
be prevented.