Store authenticated TLS Client Certificate in Event header
If a TLS connection has resulted in an authenticated TLS Client
Cert (either from client-auth "want" or "need"), then store the
certificate in a Flume event header called "ssl-cert".

The certificate is in DER format, encoded with Base 64. (This is
similar to PEM, but without the headers or line splits.)

Any other value in this header will be replaced, and any other use
of this header (for a non-authenticated client, or non-SSL) will
be prevented.